Move `graph.rs` to `infra/` in `oct-cloud`
It'll be then split into multiple files for:
* Resources
* Graph functions
* State
