+++ crates/oct-cloud/src/infra/mod.rs
@@ -1,0 +1,1 @@
+pub mod graph;
